projects/ad3552r: use the new example infrastructure
Make the necessary changes to use the new example infrastructure.
This project in particular lacked a builds.json so that is fixed as
well.
