+++ projects/ad3552r_fmcz/Makefile
@@ -1,11 +1,11 @@
 # Uncomment to use the desired platform
 # PLATFORM = xilinx
 
-# Select the example you want to enable by choosing y for enabling and n for disabling
-GENERIC_SPI_EXAMPLE = n
-AXI_QSPI_EXAMPLE = y
+EXAMPLE := axi_qspi
 
 include ../../tools/scripts/generic_variables.mk
+
+include ../../tools/scripts/examples.mk
 
 include src.mk
