+++ projects/ad3552r_fmcz/src.mk
@@ -1,27 +1,5 @@
-include $(PROJECT)/src/platform/$(PLATFORM)/platform_src.mk
-include $(PROJECT)/src/examples/examples_src.mk
-
-ifeq (y,$(strip $(IIOD)))
-LIBRARIES += iio
-SRC_DIRS += $(NO-OS)/iio/iio_app
-endif
-
-SRCS += $(PROJECT)/src/platform/$(PLATFORM)/main.c
-
-INCS += $(PROJECT)/src/common/common_data.h
-SRCS += $(PROJECT)/src/common/common_data.c
-
-INCS += $(PROJECT)/src/platform/platform_includes.h
-
-INCS += $(PROJECT)/src/platform/$(PLATFORM)/parameters.h
-
 SRCS += $(DRIVERS)/dac/ad3552r/ad3552r.c
 INCS += $(DRIVERS)/dac/ad3552r/ad3552r.h
-
-ifeq (y,$(strip $(IIOD)))
-SRCS += $(DRIVERS)/dac/ad3552r/iio_ad3552r.c
-INCS += $(DRIVERS)/dac/ad3552r/iio_ad3552r.h
-endif
 
 SRCS += $(NO-OS)/util/no_os_util.c \
         $(NO-OS)/util/no_os_alloc.c \
@@ -30,17 +8,12 @@ SRCS += $(NO-OS)/util/no_os_util.c \
         $(NO-OS)/util/no_os_crc8.c \
 	$(DRIVERS)/api/no_os_spi.c \
         $(DRIVERS)/api/no_os_gpio.c \
-        $(NO-OS)/util/no_os_util.c
-ifeq (y,$(strip $(IIOD)))
-SRCS += $(DRIVERS)/api/no_os_uart.c \
+        $(NO-OS)/util/no_os_util.c \
+	$(DRIVERS)/api/no_os_uart.c \
 	$(DRIVERS)/api/no_os_irq.c \
 	$(NO-OS)/util/no_os_lf256fifo.c \
 	$(NO-OS)/util/no_os_list.c \
-	$(NO-OS)/util/no_os_fifo.c \
-	$(PLATFORM_DRIVERS)/$(PLATFORM)_uart.c \
-        $(PLATFORM_DRIVERS)/$(PLATFORM)_irq.c \
-        $(DRIVERS)/axi_core/iio_axi_dac/iio_axi_dac.c
-endif
+	$(NO-OS)/util/no_os_fifo.c
 
 INCS += $(INCLUDE)/no_os_timer.h \
 	$(INCLUDE)/no_os_spi.h \
@@ -53,7 +26,12 @@ INCS += $(INCLUDE)/no_os_timer.h \
 	$(INCLUDE)/no_os_util.h \
 	$(INCLUDE)/no_os_alloc.h \
 	$(INCLUDE)/no_os_mutex.h \
-	$(INCLUDE)/no_os_axi_io.h
+	$(INCLUDE)/no_os_axi_io.h \
+	$(INCLUDE)/no_os_uart.h \
+	$(INCLUDE)/no_os_lf256fifo.h \
+	$(INCLUDE)/no_os_irq.h \
+	$(INCLUDE)/no_os_list.h \
+	$(INCLUDE)/no_os_fifo.h
 
 SRCS += $(DRIVERS)/axi_core/axi_dac_core/axi_dac_core.c \
 	$(DRIVERS)/axi_core/clk_axi_clkgen/clk_axi_clkgen.c \
@@ -63,21 +41,9 @@ INCS += $(DRIVERS)/axi_core/axi_dac_core/axi_dac_core.h \
 	$(DRIVERS)/axi_core/clk_axi_clkgen/clk_axi_clkgen.h \
 	$(DRIVERS)/axi_core/axi_dmac/axi_dmac.h
 
-INCS += $(PLATFORM_DRIVERS)/xilinx_gpio.h \
-	$(PLATFORM_DRIVERS)/xilinx_spi.h
-
-SRCS += $(PLATFORM_DRIVERS)/xilinx_axi_io.c \
-	$(PLATFORM_DRIVERS)/xilinx_gpio.c \
-	$(PLATFORM_DRIVERS)/xilinx_delay.c \
-	$(PLATFORM_DRIVERS)/xilinx_spi.c
-
 ifeq (y,$(strip $(IIOD)))
-INCS += $(INCLUDE)/no_os_uart.h \
-	$(INCLUDE)/no_os_lf256fifo.h \
-	$(INCLUDE)/no_os_irq.h \
-	$(INCLUDE)/no_os_list.h \
-	$(INCLUDE)/no_os_fifo.h \
-	$(PLATFORM_DRIVERS)/$(PLATFORM)_uart.h \
-	$(PLATFORM_DRIVERS)/$(PLATFORM)_irq.h \
-	$(DRIVERS)/axi_core/iio_axi_dac/iio_axi_dac.h
+SRCS += $(DRIVERS)/axi_core/iio_axi_dac/iio_axi_dac.c
+INCS += $(DRIVERS)/axi_core/iio_axi_dac/iio_axi_dac.h
+SRCS += $(DRIVERS)/dac/ad3552r/iio_ad3552r.c
+INCS += $(DRIVERS)/dac/ad3552r/iio_ad3552r.h
 endif
