+++ projects/ad3552r_fmcz/src/common/common_data.c
@@ -102,52 +102,3 @@ struct no_os_uart_init_param uart_init_param = {
 	.platform_ops = &xil_uart_ops,
 };
 #endif
-
-struct ad3552r_init_param default_ad3552r_param = {
-	.chip_id = AD3552R_ID,
-	.spi_param = {
-		.device_id = SPI_DEVICE_ID,
-		.chip_select = 0,
-		.mode = NO_OS_SPI_MODE_0,
-#ifdef GENERIC_SPI_EXAMPLE
-		.max_speed_hz = 20000000,
-#else
-		.max_speed_hz = 66000000,
-#endif
-		.bit_order = NO_OS_SPI_BIT_ORDER_MSB_FIRST,
-		.platform_ops = &xil_spi_ops,
-		.extra = SPI_EXTRA
-	},
-	.ldac_gpio_param_optional = &gpio_ldac_param,
-	.reset_gpio_param_optional = &gpio_reset_param,
-	.sdo_drive_strength = 1,
-	.channels = {
-		[0] = {
-			.en = 1,
-			.range = AD3552R_CH_OUTPUT_RANGE_NEG_10__10V,
-#ifdef AXI_QSPI_EXAMPLE
-			.fast_en = 1,
-#endif
-		},
-		[1] = {
-			.en = 1,
-			.range = AD3552R_CH_OUTPUT_RANGE_NEG_10__10V,
-#ifdef AXI_QSPI_EXAMPLE
-			.fast_en = 1,
-#endif
-		}
-	},
-	.crc_en = 0,
-	/*
-	 * Zed board requires this option, spi instruction/addr + data
-	 * must be sent in a single transfer.
-	 */
-	.single_transfer = 1,
-#ifdef AXI_QSPI_EXAMPLE
-	.axi_qspi_controller = 1,
-	.axi_clkgen_rate = 133000000,
-	.ad3552r_core_ip = &ad3552r_core_ip,
-	.dmac_ip = &dmac_ip,
-	.clkgen_ip = &clkgen_ip,
-#endif
-};
